# internal/models/responses.go
package models

type ListMetadata struct {
	Total int `json:"total"`
	Limit int `json:"limit,omitempty"`
}

type PostalCodeDetails struct {
	PostalCode     string `json:"postalCode"`
	Settlement     string `json:"settlement"`
	SettlementType string `json:"settlementType"`
	Municipality   string `json:"municipality"`
	State          string `json:"state"`
	City           string `json:"city"`
}

type SettlementSearchResult struct {
	Name           string `json:"name"`
	PostalCode     string `json:"postalCode"`
	SettlementType string `json:"settlementType"`
	Municipality   string `json:"municipality"`
	State          string `json:"state"`
	City           string `json:"city"`
}

type SettlementTypeResponse struct {
	ID   string `json:"id"`
	Name string `json:"name"`
	Slug string `json:"slug"`
}

type StatesResponse struct {
	Data []*State `json:"data"`
}

type MunicipalitiesResponse struct {
	Data []*Municipality `json:"data"`
}

type PostalCodeDetailsResponse struct {
	Data []*PostalCodeDetails `json:"data"`
}